Changes in RCON, keeping connections alive.

+++ src/UNObot/Services/EmbedDisplayService.cs
@@ -25,7 +25,7 @@ namespace UNObot.Services
 
             uint cardColor = card.Color switch
             {
-                "Red" => 0xFF0000,
+                "Red" => (uint) 0xFF0000,
                 "Blue" => 0x0000FF,
                 "Yellow" => 0xFFFF00,
                 _ => 0x00FF00,
